removed docker install logic
